+++ index.js
@@ -8,9 +8,8 @@ var Node = require('./lib/components/Node').default
 var NodeOptions = require('./lib/components/NodeOptions').default
 var Card = require('./lib/components/Card').default
 var ProductImage = require('./lib/components/ProductImage').default
-var ProductDescription = require('./lib/components/ProductDescription').default
-var ProductCard = require('./lib/components/ProductCard').default
-var ResponsiveComponentExample = require('./lib/components/ResponsiveComponentExample').default
+var ExpansionIndicator = require('./lib/components/ExpansionIndicator').default
+var ToggleableCategory = require('./lib/components/ToggleableCategory').default
 
 module.exports = {
   Button: Button,
@@ -23,7 +22,6 @@ module.exports = {
   NodeOptions: NodeOptions,
   Card: Card,
   ProductImage: ProductImage,
-  ProductDescription: ProductDescription,
-  ProductCard: ProductCard,
-  ResponsiveComponentExample: ResponsiveComponentExample,
+  ExpansionIndicator: ExpansionIndicator,
+  ToggleableCategory: ToggleableCategory,
 }