+++ src/components/ProductCard/index.js
@@ -1,3 +1,0 @@
-import ProductCard from './ProductCard.jsx'
-
-export default ProductCard