+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
-## 1.3.4
-- [fix] Fix 'ProductCard' responsiveness
+## 1.3.2
+- [new] `ExpansionIndicator` and `ToggleableCategory`
+- [removed] `ProductCard` and `ProductDescription`
 
 ## 1.3.1
+- [fix] Fix 'ProductCard' responsiveness
 - [fix] Fix react-container-query dependency
 
 ## 1.3.0